requirements.txt
Pillow
Flask
gunicorn
requests[security]
numpy
json
urllib2